+++ CHANGELOG.md
@@ -8,8 +8,11 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ### Added
 - elastic_class.ElasticSearch: Added is_active method.
+- elastic_class.ElasticSearch: Added get_shards method.
 
 ### Deprecated
+- elastic_class.is_active function.
+- elastic_class.get_shards function.
 - elastic.py: The port and scheme entries in the configuration file.